leave last failed build in place

+++ src/delete-failed-builds-in-all-jobs.groovy
@@ -1,8 +1,19 @@
 Jenkins.instance.getAllItems(AbstractProject.class).each {job ->
-    println job.fullName;
+    count = 0
+
     job.getBuilds().each {build ->
         if (build.result == Result.FAILURE) {
-            build.delete()
+            count++
+        }
+    }
+
+    println(job.fullName + " " + count)
+
+    if (count > 1) {
+        job.getBuilds().each {build ->
+            if (build.result == Result.FAILURE) {
+                build.delete()
+            }
         }
     }
 }
